Separate BaseSelect into smaller components
