+++ frontend/src/ui/Select/types.ts
@@ -1,0 +1,6 @@
+export type Option = {
+  value: string
+  label: string
+}
+
+export type Options = Array<string> | Array<Option>
